combining openai macros and functions to becca
