+++ main folder/api_file.py
@@ -1,2 +1,2 @@
 # TODO: add to .gitignore later
-API_KEY = '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'
+API_KEY = ''
